feat: force direct video download
