+++ client/src/utils/download.ts
@@ -1,0 +1,15 @@
+export async function forceDownload(url: string, filename: string) {
+  const res = await fetch(url, {mode: 'cors'});
+  if (!res.ok) {
+    throw new Error(`HTTP ${res.status}`);
+  }
+  const blob = await res.blob();
+  const objectUrl = URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = objectUrl;
+  a.download = filename;
+  document.body.appendChild(a);
+  a.click();
+  a.remove();
+  URL.revokeObjectURL(objectUrl);
+}
